Create SELU Many Layers
I am new to GitHub, so I may be doing this incorrectly. I wanted to share some code to compare ReLU vs ELU vs SELU on the MNIST data (Chapter 10 and 11). The code isn't interesting (it's all from the book) but I think the results are interesting. With 50 hidden layers, ReLU and ELU can learn if they use He initialization, but the performance is poor; however, SELU learns as well as a shallow network even without He initialization.

The book is amazing! The effort the author put in to help readers learn and understand is obvious, and very much appreciated! I am looking forward to future editions!