add ml-project-checklist (Appendix B of the book)
I am currently reading your book (physical copy) and found Appendix B super useful to rehearse the steps of a complete Machine Learning project.  

I thus produced an online copy of the appendix for myself for it to be easily accessible even without the book, and I thought it could be useful to other people.